move visGeometry and Orchestrator
