+++ src/index.ts
@@ -1,14 +1,20 @@
-import Viewport, { RenderStyle } from "./viewport";
-import Orchestrator from "./orchestrator";
-import SimulariumController from "./controller";
+import Viewport from "./viewport";
+
 export type {
     SelectionStateInfo,
     UIDisplayData,
     SimulariumFileFormat,
     VisDataFrame,
 } from "./simularium";
-export { Orchestrator, RenderStyle, SimulariumController };
-export { RemoteSimulator, ErrorLevel, FrontEndError } from "./simularium";
+
+export { RenderStyle } from "./viewport";
+export { SimulariumController } from "./controller";
+export {
+    Orchestrator,
+    RemoteSimulator,
+    ErrorLevel,
+    FrontEndError,
+} from "./simularium";
 export { compareTimes } from "./util";
 
 export default Viewport;
