+++ src/simularium/index.ts
@@ -9,7 +9,7 @@ export type {
 export type { SelectionStateInfo, UIDisplayData } from "./SelectionInterface";
 export { ErrorLevel, FrontEndError } from "./FrontEndError";
 export { RemoteSimulator, NetMessageEnum } from "./RemoteSimulator";
-export { VisGeometry, RenderStyle, NO_AGENT } from "./VisGeometry";
 export { VisData } from "./VisData";
 export { ThreadUtil } from "./ThreadUtil";
 export { SelectionInterface } from "./SelectionInterface";
+export { Orchestrator } from "./orchestrator";
