+++ src/test/GeometryStore.test.ts
@@ -3,13 +3,11 @@ import { Mesh } from "three";
 import {
     InstancedMesh,
     InstanceType,
-} from "../simularium/VisGeometry/rendering/InstancedMesh";
-import VisAgent from "../simularium/VisGeometry/VisAgent";
-import GeometryStore, {
-    DEFAULT_MESH_NAME,
-} from "../simularium/VisGeometry/GeometryStore";
-import { GeometryDisplayType } from "../simularium/VisGeometry/types";
-import PDBModel from "../simularium/VisGeometry/PDBModel";
+} from "../visGeometry/rendering/InstancedMesh";
+import VisAgent from "../visGeometry/VisAgent";
+import GeometryStore, { DEFAULT_MESH_NAME } from "../visGeometry/GeometryStore";
+import { GeometryDisplayType } from "../visGeometry/types";
+import PDBModel from "../visGeometry/PDBModel";
 
 describe("GeometryStore module", () => {
     test("it creates a registry with a single mesh", () => {
