+++ src/visGeometry/KMeansWorker.ts
