+++ src/visGeometry/agentPath.ts
