+++ src/visGeometry/rendering/ContourPass.ts
