+++ src/visGeometry/rendering/GaussianBlur.ts
