+++ src/visGeometry/rendering/HitTestHelper.ts
