+++ src/visGeometry/rendering/InstancedFiber.ts
