+++ src/visGeometry/rendering/InstancedFiberShader.ts
