+++ src/visGeometry/rendering/KMeans3d.ts
