+++ src/visGeometry/rendering/LegacyRenderer.ts
