+++ src/visGeometry/rendering/SimulariumRenderer.ts
