+++ src/visGeometry/types.ts
